+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Enable stub mode within `TYPE_CHECKING` branches (#702)
 - Infer from overloads - add default value in impl (#697)
 - Warn for missing returns with explicit `Any` return types (#715)
+- Added `--helpful-string-allow-none` to allow `None` with `helpful-string` (#717)
+- Enabled `helpful-string` by default (#717)
 ### Fixes
 - positional arguments on overloads break super (#697)
 - positional arguments on overloads duplicate unions (#697)